Model pattern off of usenode
