+++ Tasks/UseDotNetV1/installer.ts
@@ -1,0 +1,142 @@
+import * as tl from 'azure-pipelines-task-lib/task';
+import * as trm from 'azure-pipelines-task-lib/toolrunner';
+import * as toolLib from 'azure-pipelines-tool-lib/tool';
+import { DotNetCoreReleaseFetcher } from "./releasesfetcher";
+import * as utilities from "./utilities";
+
+export class DotnetCoreInstaller {
+    constructor(packageType, version) {
+        this.packageType = packageType;
+        if (!toolLib.isExplicitVersion(version)) {
+            throw tl.loc("ImplicitVersionNotSupported", version);
+        }
+        this.version = version;
+        this.cachedToolName = this.packageType === 'runtime' ? 'dncr' : 'dncs';;
+    }
+
+    public async install() {
+        // Check cache
+        let toolPath: string;
+        let osSuffixes = this.detectMachineOS();
+        let parts = osSuffixes[0].split("-");
+        this.arch = parts.length > 1 ? parts[1] : "x64";
+        toolPath = this.getLocalTool();
+
+        if (!toolPath) {
+            // download, extract, cache
+            console.log(tl.loc("InstallingAfresh"));
+            console.log(tl.loc("GettingDownloadUrl", this.packageType, this.version));
+            let downloadUrls = await DotNetCoreReleaseFetcher.getDownloadUrls(osSuffixes, this.version, this.packageType);
+            toolPath = await this.downloadAndInstall(downloadUrls);
+        } else {
+            console.log(tl.loc("UsingCachedTool", toolPath));
+        }
+
+        // Prepend the tools path. instructs the agent to prepend for future tasks
+        toolLib.prependPath(toolPath);
+
+        try {
+            let globalToolPath: string = "";
+            if (tl.osType().match(/^Win/)) {
+                globalToolPath = path.join(process.env.USERPROFILE, ".dotnet\\tools");
+            } else {
+                globalToolPath = path.join(process.env.HOME, ".dotnet/tools");
+            }
+
+            console.log(tl.loc("PrependGlobalToolPath"));
+            tl.mkdirP(globalToolPath);
+            toolLib.prependPath(globalToolPath);
+        } catch (error) {
+            //nop
+        }
+
+        // Set DOTNET_ROOT for dotnet core Apphost to find runtime since it is installed to a non well-known location.
+        tl.setVariable('DOTNET_ROOT', toolPath);
+    }
+
+    private getLocalTool(): string {
+        console.log(tl.loc("CheckingToolCache"));
+        return toolLib.findLocalTool(this.cachedToolName, this.version, this.arch);
+    }
+
+    private detectMachineOS(): string[] {
+        let osSuffix = [];
+        let scriptRunner: trm.ToolRunner;
+
+        if (tl.osType().match(/^Win/)) {
+            let escapedScript = path.join(utilities.getCurrentDir(), 'externals', 'get-os-platform.ps1').replace(/'/g, "''");
+            let command = `& '${escapedScript}'`
+
+            let powershellPath = tl.which('powershell', true);
+            scriptRunner = tl.tool(powershellPath)
+                .line('-NoLogo -Sta -NoProfile -NonInteractive -ExecutionPolicy Unrestricted -Command')
+                .arg(command);
+        }
+        else {
+            let scriptPath = path.join(utilities.getCurrentDir(), 'externals', 'get-os-distro.sh');
+            utilities.setFileAttribute(scriptPath, "777");
+
+            scriptRunner = tl.tool(tl.which(scriptPath, true));
+        }
+
+        let result: trm.IExecSyncResult = scriptRunner.execSync();
+
+        if (result.code != 0) {
+            throw tl.loc("getMachinePlatformFailed", result.error ? result.error.message : result.stderr);
+        }
+
+        let output: string = result.stdout;
+
+        let index;
+        if ((index = output.indexOf("Primary:")) >= 0) {
+            let primary = output.substr(index + "Primary:".length).split(os.EOL)[0];
+            osSuffix.push(primary);
+            console.log(tl.loc("PrimaryPlatform", primary));
+        }
+
+        if ((index = output.indexOf("Legacy:")) >= 0) {
+            let legacy = output.substr(index + "Legacy:".length).split(os.EOL)[0];
+            osSuffix.push(legacy);
+            console.log(tl.loc("LegacyPlatform", legacy));
+        }
+
+        if (osSuffix.length == 0) {
+            throw tl.loc("CouldNotDetectPlatform");
+        }
+
+        return osSuffix;
+    }
+
+    private async downloadAndInstall(downloadUrls: string[]) {
+        let downloaded = false;
+        let downloadPath = "";
+        for (const url of downloadUrls) {
+            try {
+                downloadPath = await toolLib.downloadTool(url);
+                downloaded = true;
+                break;
+            } catch (error) {
+                tl.warning(tl.loc("CouldNotDownload", url, JSON.stringify(error)));
+            }
+        }
+
+        if (!downloaded) {
+            throw tl.loc("FailedToDownloadPackage");
+        }
+
+        // extract
+        console.log(tl.loc("ExtractingPackage", downloadPath));
+        let extPath: string = tl.osType().match(/^Win/) ? await toolLib.extractZip(downloadPath) : await toolLib.extractTar(downloadPath);
+
+        // cache tool
+        console.log(tl.loc("CachingTool"));
+        let cachedDir = await toolLib.cacheDir(extPath, this.cachedToolName, this.version, this.arch);
+        console.log(tl.loc("SuccessfullyInstalled", this.packageType, this.version));
+        return cachedDir;
+    }
+
+    private packageType: string;
+    private version: string;
+    private cachedToolName: string;
+    private arch: string;
+}
